Float -> Double
Float-datatyypin desimaalikäsittely on rikki Koskessa, esim. 4.6 muuttuu 4.59999991223

+++ src/main/scala/fi/oph/koski/schema/Ammatillinen.scala
@@ -245,7 +245,7 @@ case class AmmatillisenTutkinnonSuoritus(
   @MinValue(1)
   @MaxValue(5)
   @Scale(2)
-  keskiarvo: Option[Float] = None
+  keskiarvo: Option[Double] = None
 ) extends AmmatillisenTutkinnonOsittainenTaiKokoSuoritus with Todistus
 
 @ReadFlattened
